requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
